Run 4 git-blame processes in parallel
Start a couple of blame processes, and read all their results.  Repeat this
process until no more hunks are left.  This speeds up the blaming part on
large repositories/changesets.

The value of 4 happened to work best on my machine on a very quick benchmark
run on an Intel(R) Core(TM) i5-6300U CPU @ 2.40GHz with four CPU cores.
I made 1000 random changes to files touched in HEAD~1000.. in linux. Running
"git autofixup HEAD~1000" produced 164 fixup commits.
The runtime improved from 74 seconds to 51 seconds. A significant share of
the time is spent on creating the commits, which is the same in both cases.

Maybe the number of git-blame processes should be chosen dynamically based
on the number of CPUs. I have not needed that so far.  Also my benchmark
is a bit unrealistic.